add initial version of network processes detection
Implement core functions to query the kernel for processes with active
TCP/UDP connections (IPv4/IPv6). Uses netlink SOCK_DIAG interface to
retrieve socket information and associate it with process IDs by
scanning /proc.

This is a WIP, still trying to understand how everything related to
this works. Lots of AI code in here that might change (and I already
removed a lot).